Refactoring examples by interpreters
